rename error to errors for consistency with go language
